Added AddInstrumentation test.
